fix 2j overflow in UTA
git-svn-id: file:///home/fnevgeny/svn/fac/trunk@639 ad7935cb-1c1c-0410-8b58-bafadb17b910

+++ fac/faclib/crm.h
@@ -54,7 +54,7 @@ typedef struct _ION_ {
   int nlevels;
   LBLOCK **iblock;
   int *ilev;
-  short *j;
+  int *j;
   short *vnl;
   short *ibase;
   double *energy;
